@@ -4,6 +4,7 @@ import (
 	"context"
 	errors2 "errors"
 	"fmt"
+	"log/slog"
 	"net/http"
 	"net/url"
 	"path"
@@ -11,17 +12,14 @@ import (
 	"sync"
 	"time"
 
-	"encoding/json"
-
 	"github.com/micro/micro/v3/service/context/metadata"
 	"github.com/micro/micro/v3/service/errors"
-	"github.com/rs/zerolog"
-
 	auth "github.com/webitel/chat_manager/api/proto/auth"
 	gate "github.com/webitel/chat_manager/api/proto/bot"
 	chat "github.com/webitel/chat_manager/api/proto/chat"
 	"github.com/webitel/chat_manager/app"
 	access "github.com/webitel/chat_manager/auth"
+	wlog "github.com/webitel/chat_manager/log"
 )
 
 const ExternalChatPropertyName = "externalChatID"
@@ -30,7 +28,7 @@ const ExternalChatPropertyName = "externalChatID"
 type Gateway struct {
 	// identity
 	*Bot // *chat.Profile
-	Log  *zerolog.Logger
+	Log  *slog.Logger
 	// Template of .Bot.Updates; Compiled
 	Template *Template
 	// communication
@@ -115,16 +113,14 @@ func (c *Gateway) Register(ctx context.Context, force bool) error {
 		err = c.External.Register(ctx, linkURI)
 	}
 
-	var event *zerolog.Event
-
 	if err == nil {
-		event = c.Log.Info()
+		c.Log.Info("[ GATE::REGISTER ]")
 	} else {
 		_ = c.Remove()
-		event = c.Log.Error().Err(err)
+		c.Log.Error("[ GATE::REGISTER ]",
+			slog.Any("error", err),
+		)
 	}
-
-	event.Msg("REGISTER")
 
 	return err
 }
@@ -194,20 +190,21 @@ func (c *Gateway) Deregister(ctx context.Context) error {
 	link := strings.TrimRight(srv.URL, "/") + uri
 
 	var (
-		event *zerolog.Event
 		// PERFORM: DEREGISTER
 		err = c.External.Deregister(ctx)
 	)
 
-	if err == nil {
-		event = c.Log.Warn()
-	} else {
-		event = c.Log.Error().Err(err)
-	}
+	log := c.Log.With(
+		slog.String("link", link),
+	)
 
-	event.
-		Str("link", link).
-		Msg("DEREGISTER")
+	if err == nil {
+		log.Warn("DEREGISTER")
+	} else {
+		log.Error("DEREGISTER",
+			slog.Any("error", err),
+		)
+	}
 
 	return err
 }
@@ -231,7 +228,7 @@ func (c *Gateway) Remove() bool {
 	srv.indexMx.Unlock() // -RW
 
 	if ok {
-		c.Log.Warn().Msg("DELETED")
+		c.Log.Warn("[ GATE:EVICTED ]")
 	}
 
 	// var event *zerolog.Event
@@ -420,7 +417,9 @@ func (c *Gateway) GetChannel(ctx context.Context, chatID string, contact *Accoun
 		chat, err := c.Internal.Client.CheckSession(ctx, &lookup)
 
 		if err != nil {
-			c.Log.Error().Err(err).Msg("Failed to lookup chat channel")
+			c.Log.Error("Failed to lookup chat channel",
+				slog.Any("error", err),
+			)
 			return nil, err
 		}
 
@@ -444,14 +443,14 @@ func (c *Gateway) GetChannel(ctx context.Context, chatID string, contact *Accoun
 				Gateway:    c, // .profile.id
 				Properties: chat.Properties,
 
-				Log: c.Log.With().
-					Str("chat-id", chatID).
-					Str("username", contact.Username).
-
-					// Str("session-id", c.SessionID). // UNKNOWN
-					Str("channel-id", chat.ChannelId).
-					Int64("contact-id", chat.ClientId).
-					Logger(),
+				Log: c.Log.With(
+					"chat", slog.GroupValue(
+						slog.String("user", contact.Channel+":"+chatID),
+						slog.String("title", contact.Username),
+						slog.Int64("uid", chat.ClientId),
+						slog.String("id", chat.ChannelId),
+					),
+				),
 			}
 
 			c.Lock() // +RW
@@ -459,7 +458,7 @@ func (c *Gateway) GetChannel(ctx context.Context, chatID string, contact *Accoun
 			c.internal[channel.Account.ID] = channel // [channel.ContactID] = channel
 			c.Unlock()                               // -RW
 
-			channel.Log.Info().Msg("RECOVER")
+			channel.Log.Info("[ CHAT::RECOVER ]")
 
 		} else {
 
@@ -471,7 +470,7 @@ func (c *Gateway) GetChannel(ctx context.Context, chatID string, contact *Accoun
 			// NO Channel FOUND !
 			// CHECK: Can we accept NEW one ?
 			if !c.Bot.GetEnabled() {
-				c.Log.Warn().Msg("DISABLED")
+				c.Log.Warn("[ GATE::DISABLED ]")
 				return nil, errors.New(
 					"chat.bot.channel.disabled",
 					"chat: bot is disabled",
@@ -496,13 +495,13 @@ func (c *Gateway) GetChannel(ctx context.Context, chatID string, contact *Accoun
 				// add
 				Properties: map[string]string{ExternalChatPropertyName: chatID},
 
-				Log: c.Log.With().
-					Str("chat-id", chatID).
-					Str("username", contact.Username).
-
-					// Str("session-id", c.SessionID). // UNKNOWN
-					Int64("contact-id", chat.ClientId).
-					Logger(),
+				Log: c.Log.With(
+					"chat", slog.GroupValue(
+						slog.String("user", contact.Channel+":"+chatID),
+						slog.String("title", contact.Username),
+						slog.Int64("uid", chat.ClientId),
+					),
+				),
 			}
 
 			c.Lock() // +RW
@@ -510,7 +509,7 @@ func (c *Gateway) GetChannel(ctx context.Context, chatID string, contact *Accoun
 			c.internal[channel.Account.ID] = channel // [channel.ContactID] = channel
 			c.Unlock()                               // -RW
 
-			channel.Log.Info().Msg("PREPARE")
+			channel.Log.Info("[ CHAT::MAKE ]")
 
 			// // .IsNew() == true
 			// return channel, nil
@@ -541,7 +540,10 @@ func (c *Gateway) GetChannel(ctx context.Context, chatID string, contact *Accoun
 				var err error
 				channel.Title, err = c.Template.MessageText("title", customer)
 				if err != nil {
-					channel.Log.Warn().Err(err).Msg("BOT.onContactUpdate")
+					channel.Log.Warn(
+						"chat.onContactUpdate",
+						slog.Any("error", err),
+					)
 					err = nil
 				}
 			}
@@ -556,15 +558,13 @@ func (c *Gateway) GetChannel(ctx context.Context, chatID string, contact *Accoun
 
 			if customer.Channel != "whatsapp" {
 				// panic("BOT.onContactUpdate: client.external_id changed; client.channel(" + customer.Channel + ") not supported")
-				channel.Log.Warn().
-					Str("error", customer.Channel+": no support; accept: whatsapp").
-					// Str("chat-id", chatID). // OLD client.external_id
-					// Int64("contact-id", customer.ID). // BOT.client.id
-					// Str("channel", c.GetProvider()). // BOT.provider(channel-type)
-					Str("new-contact", customer.Channel). // BOT.client.type
-					Str("new-chat-id", newChatId).        // NEW client.external_id
-					Str("new-title", channel.Title).
-					Msg("BOT.onContactUpdate")
+				channel.Log.Warn(
+					"chat.onContactUpdate",
+					slog.String("error", customer.Channel+": no support; accept: whatsapp"),
+					slog.String("new.title", channel.Title),      // NEW client.external_id
+					slog.String("new.contact", customer.Channel), // BOT.client.type
+				)
+
 				return channel, nil
 			}
 
@@ -591,13 +591,14 @@ func (c *Gateway) GetChannel(ctx context.Context, chatID string, contact *Accoun
 
 		if updateChatId || updateChatTitle {
 			// Update channel logger info
-			channel.Log = c.Log.With().
-				Str("chat-id", channel.ChatID).
-				Str("username", customer.DisplayName()). // Username).
-				// Str("session-id", c.SessionID). // UNKNOWN
-				Str("channel-id", channel.ChannelID).
-				Int64("contact-id", customer.ID).
-				Logger()
+			channel.Log = c.Log.With(
+				"chat", slog.GroupValue(
+					slog.String("user", customer.Channel+":"+channel.ChatID),
+					slog.String("title", customer.DisplayName()),
+					slog.Int64("uid", customer.ID),
+					slog.String("id", channel.ChannelID),
+				),
+			)
 
 			// Persist store.clients changes
 			contactName := customer.DisplayName()
@@ -620,20 +621,21 @@ func (c *Gateway) GetChannel(ctx context.Context, chatID string, contact *Accoun
 				err = fmt.Errorf("client: not found")
 			}
 
-			var log *zerolog.Event
+			log := channel.Log //.With(
+			// 	// Str("chat-id", chatID).             // NEW client.external_id
+			// 	// Int64("contact-id", customer.ID).   // BOT.client.id
+			// 	// Str("channel", c.GetProvider()).    // BOT.provider(channel-type)
+			// 	slog.String("contact-type", customer.Channel), // BOT.client.type
+			// 	slog.String("from-chat-id", chatID),           // OLD client.external_id
+			// )
 			if err == nil {
-				log = channel.Log.Info()
+				log.Info("chat.onContactUpdate")
 			} else {
-				log = channel.Log.Warn().Err(err)
+				log.Warn("chat.onContactUpdate",
+					slog.Any("error", err),
+				)
 				err = nil // LOG -and- IGNORE
 			}
-			log.
-				// Str("chat-id", chatID).             // NEW client.external_id
-				// Int64("contact-id", customer.ID).   // BOT.client.id
-				// Str("channel", c.GetProvider()).    // BOT.provider(channel-type)
-				Str("contact-type", customer.Channel). // BOT.client.type
-				Str("from-chat-id", chatID).           // OLD client.external_id
-				Msg("BOT.onContactUpdate")
 		}
 
 	}
@@ -822,7 +824,14 @@ func (c *Gateway) Send(ctx context.Context, notify *gate.SendMessageRequest) err
 	gate := recepient.Gateway
 	err = gate.External.SendNotify(ctx, &sendUpdate)
 
-	var event *zerolog.Event
+	emit := recepient.Log.With(
+		"msg", slog.GroupValue(
+			slog.Int64("id", sendUpdate.Message.GetId()),
+			slog.String("type", sendUpdate.Message.GetType()),
+			// slog.String("text", sendUpdate.Message.GetText()),
+			// slog.String("file", wlog.JsonValue(sendUpdate.Message.GetFile())),
+		),
+	)
 
 	if err == nil {
 		// FIXME: .GetChannel() does not provide full contact info on recover,
@@ -830,16 +839,12 @@ func (c *Gateway) Send(ctx context.Context, notify *gate.SendMessageRequest) err
 		// if recepient.Title == "" {
 		// 	recepient.Title = recepient.Account.DisplayName()
 		// }
-		event = recepient.Log.Debug()
+		emit.Debug("[ SEND::MESSAGE ]")
 	} else {
-		event = recepient.Log.Error().Err(err)
+		emit.Error("[ SEND::MESSAGE ]",
+			slog.Any("error", err),
+		)
 	}
-
-	event.
-		Str("send", sendUpdate.Message.GetType()). // sendUpdate.Event).
-		Str("text", sendUpdate.Message.GetText()).
-		EmbedObject(ZerologJSON("file", sendUpdate.Message.GetFile())).
-		Msg("<<<<< SEND <<<<<<")
 
 	if err != nil {
 		return err
@@ -881,32 +886,6 @@ func (c *Gateway) DeleteMessage(ctx context.Context, update *Update) error {
 	return nil // ACK(+)
 }
 
-type zerologFunc func(event *zerolog.Event)
-
-func (fn zerologFunc) MarshalZerologObject(event *zerolog.Event) {
-	fn(event)
-}
-
-func ZerologJSON(key string, obj interface{}) zerolog.LogObjectMarshaler {
-	return zerologFunc(func(event *zerolog.Event) {
-
-		if obj == nil {
-			event.Str(key, "")
-			return
-		}
-
-		data, err := json.Marshal(obj)
-
-		if err != nil {
-			event.Err(err)
-			return
-		}
-
-		event.RawJSON(key, data)
-
-	})
-}
-
 // Read notification [FROM] external: chat.provider [TO] internal: chat.server
 func (c *Gateway) Read(ctx context.Context, notify *Update) (err error) {
 
@@ -930,7 +909,9 @@ func (c *Gateway) Read(ctx context.Context, notify *Update) (err error) {
 
 	if channel.IsNew() {
 		if channel.Title, err = c.Template.MessageText("title", contact); err != nil {
-			channel.Log.Warn().Err(err).Msg("bot.updateChatTitle")
+			channel.Log.Warn("bot.updateChatTitle",
+				slog.Any("error", err),
+			)
 			err = nil
 		}
 	}
@@ -946,4 +927,8 @@ func (c *Gateway) Read(ctx context.Context, notify *Update) (err error) {
 	}
 
 	return nil // ACK(+)
+}
+
+func (c *Gateway) TraceLog(msg string, args ...any) {
+	wlog.TraceLog(c.Log, msg, args...)
 }
