@@ -5,6 +5,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"log/slog"
 	"net/http"
 	"path"
 	"strconv"
@@ -12,7 +13,6 @@ import (
 	"time"
 
 	"github.com/micro/micro/v3/service/errors"
-	"github.com/rs/zerolog/log"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
@@ -49,7 +49,7 @@ func New(agent *bot.Gateway, state bot.Provider) (bot.Provider, error) {
 	profile := agent.Bot.GetMetadata()
 	botToken, ok := profile["token"]
 	if !ok {
-		log.Error().Msg("AppToken not found")
+		agent.Log.Error("AppToken not found")
 		return nil, errors.BadRequest(
 			"chat.gateway.viber.token.required",
 			"viber: bot API token required",
@@ -178,7 +178,9 @@ func (c *Bot) Register(ctx context.Context, linkURL string) error {
 	}
 
 	if err != nil {
-		c.Gateway.Log.Err(err).Msg("viber/bot.setWebhook")
+		c.Gateway.Log.Error("viber/bot.setWebhook",
+			slog.Any("error", err),
+		)
 		return err
 	}
 
@@ -219,14 +221,6 @@ func (c *Bot) Deregister(ctx context.Context) error {
 	}
 
 	return nil
-}
-
-func contactPeer(peer *chat.Account) *chat.Account {
-	if peer.LastName == "" {
-		peer.FirstName, peer.LastName =
-			bot.FirstLastName(peer.FirstName)
-	}
-	return peer
 }
 
 func (c *Bot) SendNotify(ctx context.Context, notify *bot.Update) error {
@@ -271,14 +265,14 @@ func (c *Bot) SendNotify(ctx context.Context, notify *bot.Update) error {
 		)
 
 	case "left":
-
-		peer := contactPeer(sentMessage.LeftChatMember)
+		peer := sentMessage.LeftChatMember
 		updates := c.Gateway.Template
 		messageText, err := updates.MessageText("left", peer)
 		if err != nil {
-			c.Gateway.Log.Err(err).
-				Str("update", sentMessage.Type).
-				Msg("viber/bot.updateLeftMember")
+			c.Gateway.Log.Error("viber/bot.updateLeftMember",
+				slog.Any("error", err),
+				slog.String("update", sentMessage.Type),
+			)
 		}
 		messageText = strings.TrimSpace(
 			messageText,
@@ -290,14 +284,14 @@ func (c *Bot) SendNotify(ctx context.Context, notify *bot.Update) error {
 		sendMessage.Text(messageText)
 
 	case "joined":
-
-		peer := contactPeer(sentMessage.NewChatMembers[0])
+		peer := sentMessage.NewChatMembers[0]
 		updates := c.Gateway.Template
 		messageText, err := updates.MessageText("join", peer)
 		if err != nil {
-			c.Gateway.Log.Err(err).
-				Str("update", sentMessage.Type).
-				Msg("vider/bot.updateChatMember")
+			c.Gateway.Log.Error("viber/bot.updateChatMember",
+				slog.Any("error", err),
+				slog.String("update", sentMessage.Type),
+			)
 		}
 		messageText = strings.TrimSpace(
 			messageText,
@@ -326,9 +320,10 @@ func (c *Bot) SendNotify(ctx context.Context, notify *bot.Update) error {
 		updates := c.Gateway.Template
 		messageText, err := updates.MessageText("close", nil)
 		if err != nil {
-			c.Gateway.Log.Err(err).
-				Str("update", sentMessage.Type).
-				Msg("viber/bot.updateChatClose")
+			c.Gateway.Log.Error("viber/bot.updateChatClose",
+				slog.Any("error", err),
+				slog.String("update", sentMessage.Type),
+			)
 		}
 		messageText = strings.TrimSpace(
 			messageText,
@@ -445,7 +440,9 @@ func (c *Bot) do(r request, w interface{}) error {
 	}
 
 	if err != nil {
-		c.Gateway.Log.Err(err).Msg("viber/" + r.method() + ":result")
+		c.Gateway.Log.Error("viber/"+r.method()+":result",
+			slog.Any("error", err),
+		)
 		return err
 	}
 
@@ -474,7 +471,7 @@ func (c *Bot) WebHook(reply http.ResponseWriter, notice *http.Request) {
 	var event Update
 	err := json.NewDecoder(notice.Body).Decode(&event)
 	if err != nil {
-		c.Gateway.Log.Err(err).Msg("viber/bot.onUpdate")
+		c.Gateway.Log.Error("viber/bot.onUpdate")
 		return // (200) IGNORE
 	}
 
@@ -499,17 +496,18 @@ func (c *Bot) WebHook(reply http.ResponseWriter, notice *http.Request) {
 		updateFailMessage:
 		hook = c.onMsgStatus
 	default:
-		c.Gateway.Log.Warn().
-			Str("event", event.Type).
-			Str("error", "event: no update reaction").
-			Msg("viber/bot.onUpdate")
+		c.Gateway.Log.Warn("viber/bot.onUpdate",
+			slog.String("event", event.Type),
+			slog.String("error", "event: no update reaction"),
+		)
 		return // (200) IGNORE
 	}
 	// Handle update event
 	err = hook(ctx, &event)
 	if err != nil {
-		c.Gateway.Log.Err(err).
-			Msg("viber/bot.on" + strings.Title(event.Type))
+		c.Gateway.Log.Error("viber/bot.on"+strings.Title(event.Type),
+			slog.Any("error", err),
+		)
 		return // (200) IGNORE
 	}
 
@@ -520,11 +518,10 @@ func (c *Bot) WebHook(reply http.ResponseWriter, notice *http.Request) {
 func (c *Bot) BroadcastMessage(ctx context.Context, req *chat.BroadcastMessageRequest, rsp *chat.BroadcastMessageResponse) error {
 
 	var (
-		n       = len(req.GetPeer())
-		castErr = func(peerId string, err error) {
+		setError = func(peerId string, err error) {
 			res := rsp.GetFailure()
 			if res == nil {
-				res = make([]*chat.BroadcastPeer, 0, n)
+				res = make([]*chat.BroadcastPeer, 0, len(req.GetPeer()))
 			}
 
 			var re *status.Status
@@ -563,22 +560,35 @@ func (c *Bot) BroadcastMessage(ctx context.Context, req *chat.BroadcastMessageRe
 		}
 	)
 
-	// Prepare message text to broadcast
-	_ = cast.Text(req.GetMessage().GetText())
-	// Perform broadcast request
-	err := c.do(&cast, &res)
-	if err == nil {
-		err = res.Err()
+	// Get message params from request
+	message := req.GetMessage()
+
+	// Set text or file to message
+	switch message.GetType() {
+	case "text":
+		cast.Text(message.GetText())
+	case "file":
+		cast.Media(
+			req.Message.GetFile(),
+			req.Message.GetText(), // Max 512 characters !
+		)
 	}
 
+	// Perform broadcast request
+	err := c.do(&cast, &res)
 	if err != nil {
-		return err // ERR
+		return err
+	}
+
+	err = res.Err()
+	if err != nil {
+		return err
 	}
 
 	// Populate failed peer(s) status
 	for _, fail := range res.FailStatus {
-		castErr(fail.PeerId, fail.Err())
+		setError(fail.PeerId, fail.Err())
 	}
 
-	return nil // OK
+	return nil
 }
