@@ -5,16 +5,18 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"io"
+	"log/slog"
+	"mime/multipart"
+	"net/http"
+	"strconv"
+	"strings"
+
 	vk "github.com/SevereCloud/vksdk/v2/api"
 	"github.com/SevereCloud/vksdk/v2/object"
 	"github.com/micro/micro/v3/service/errors"
 	chat "github.com/webitel/chat_manager/api/proto/chat"
 	"github.com/webitel/chat_manager/bot"
-	"io"
-	"mime/multipart"
-	"net/http"
-	"strconv"
-	"strings"
 )
 
 const (
@@ -167,7 +169,9 @@ func (c *VKBot) Register(ctx context.Context, callbackURL string) error {
 	params := vk.Params{"group_id": c.creds.GroupId, "url": callbackURL, "title": hookName}
 	resp, err := c.BotApi.GroupsAddCallbackServer(params)
 	if err != nil {
-		c.Gateway.Log.Error().Err(err).Msg("Failed to .Register webhook")
+		c.Gateway.Log.Error("Failed to .Register webhook",
+			slog.Any("error", err),
+		)
 		return err
 	}
 	c.serverId = int64(resp.ServerID)
@@ -193,14 +197,6 @@ func (c *VKBot) Deregister(ctx context.Context) error {
 	}
 
 	return nil
-}
-
-func contactPeer(peer *chat.Account) *chat.Account {
-	if peer.LastName == "" {
-		peer.FirstName, peer.LastName =
-			bot.FirstLastName(peer.FirstName)
-	}
-	return peer
 }
 
 // SendNotify implements provider.Sender interface for VK
@@ -271,7 +267,9 @@ func (c *VKBot) WebHook(reply http.ResponseWriter, notice *http.Request) {
 	err := json.NewDecoder(notice.Body).Decode(&recvEvent)
 	if err != nil {
 		http.Error(reply, "Failed to decode vk .Update event", http.StatusBadRequest)
-		c.Log.Error().Str("error", "vk.Update: "+err.Error()).Msg("VK: UPDATE")
+		c.Log.Error("VK: UPDATE",
+			slog.String("error", "vk.Update: "+err.Error()),
+		)
 		return // 400 Bad Request
 	}
 
