@@ -4,14 +4,16 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	vk "github.com/SevereCloud/vksdk/v2/api"
-	"github.com/micro/micro/v3/service/errors"
-	"github.com/webitel/chat_manager/api/proto/chat"
-	"github.com/webitel/chat_manager/bot"
+	"log/slog"
 	"math/rand"
 	"strconv"
 	"strings"
 	"time"
+
+	vk "github.com/SevereCloud/vksdk/v2/api"
+	"github.com/micro/micro/v3/service/errors"
+	"github.com/webitel/chat_manager/api/proto/chat"
+	"github.com/webitel/chat_manager/bot"
 )
 
 type OutgoingMessage struct {
@@ -211,14 +213,16 @@ func (c *VKBot) ConvertInternalToOutcomingMessage(update *bot.Update) (*Outgoing
 			}
 			result.Attachments = append(result.Attachments, attachment)
 		}
+
 	case "joined":
-		peer := contactPeer(message.NewChatMembers[0])
+		peer := message.NewChatMembers[0]
 		updates := c.Gateway.Template
 		text, err := updates.MessageText("join", peer)
 		if err != nil {
-			c.Gateway.Log.Err(err).
-				Str("update", message.Type).
-				Msg("vk/bot.updateChatMember")
+			c.Gateway.Log.Error("vk/bot.updateChatMember",
+				slog.Any("error", err),
+				slog.String("update", message.Type),
+			)
 		}
 		//if text != "" {
 		// format new message to the engine for saving it in the DB as operator message [WTEL-4695]
@@ -241,28 +245,32 @@ func (c *VKBot) ConvertInternalToOutcomingMessage(update *bot.Update) (*Outgoing
 		//}
 
 	case "left":
-		peer := contactPeer(message.LeftChatMember)
+		peer := message.LeftChatMember
 		updates := c.Gateway.Template
 		text, err := updates.MessageText("left", peer)
 		if err != nil {
-			c.Gateway.Log.Err(err).
-				Str("update", message.Type).
-				Msg("vk/bot.updateChatMember")
+			c.Gateway.Log.Error("vk/bot.updateChatMember",
+				slog.Any("error", err),
+				slog.String("update", message.Type),
+			)
 		}
 		//if text != "" {
 		result.Text = text
 		//}
+
 	case "closed":
 		updates := c.Gateway.Template
 		text, err := updates.MessageText("close", nil)
 		if err != nil {
-			c.Gateway.Log.Err(err).
-				Str("update", message.Type).
-				Msg("vk/bot.updateChatMember")
+			c.Gateway.Log.Error("vk/bot.updateChatMember",
+				slog.Any("error", err),
+				slog.String("update", message.Type),
+			)
 		}
 		//if text != "" {
 		result.Text = text
 		//}
+
 	default:
 		messageText := strings.TrimSpace(
 			message.GetText(),
